+++ server/src/grpc/storage_service/mod.rs
@@ -441,133 +441,3 @@ impl<Q: QueryExecutor + 'static> StorageService for StorageServiceImpl<Q> {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use ceresdbproto::storage::{value, Field, FieldGroup, Tag, Value, WriteSeriesEntry};
-    use cluster::config::SchemaConfig;
-    use common_types::datum::DatumKind;
-
-    use super::*;
-
-    const TAG1: &str = "host";
-    const TAG2: &str = "idc";
-    const FIELD1: &str = "cpu";
-    const FIELD2: &str = "memory";
-    const FIELD3: &str = "log";
-    const FIELD4: &str = "ping_ok";
-    const TABLE: &str = "pod_system_table";
-    const TIMESTAMP_COLUMN: &str = "custom_timestamp";
-
-    fn make_tag(name_index: u32, val: &str) -> Tag {
-        Tag {
-            name_index,
-            value: Some(Value {
-                value: Some(value::Value::StringValue(val.to_string())),
-            }),
-        }
-    }
-
-    fn make_field(name_index: u32, val: value::Value) -> Field {
-        Field {
-            name_index,
-            value: Some(Value { value: Some(val) }),
-        }
-    }
-
-    fn generate_write_table_request() -> WriteTableRequest {
-        let tag1 = make_tag(0, "test.host");
-        let tag2 = make_tag(1, "test.idc");
-        let tags = vec![tag1, tag2];
-
-        let field1 = make_field(0, value::Value::Float64Value(100.0));
-        let field2 = make_field(1, value::Value::Float64Value(1024.0));
-        let field3 = make_field(2, value::Value::StringValue("test log".to_string()));
-        let field4 = make_field(3, value::Value::BoolValue(true));
-
-        let field_group1 = FieldGroup {
-            timestamp: 1000,
-            fields: vec![field1.clone(), field4],
-        };
-        let field_group2 = FieldGroup {
-            timestamp: 2000,
-            fields: vec![field1, field2],
-        };
-        let field_group3 = FieldGroup {
-            timestamp: 3000,
-            fields: vec![field3],
-        };
-
-        let write_entry = WriteSeriesEntry {
-            tags,
-            field_groups: vec![field_group1, field_group2, field_group3],
-        };
-
-        let tag_names = vec![TAG1.to_string(), TAG2.to_string()];
-        let field_names = vec![
-            FIELD1.to_string(),
-            FIELD2.to_string(),
-            FIELD3.to_string(),
-            FIELD4.to_string(),
-        ];
-
-        WriteTableRequest {
-            table: TABLE.to_string(),
-            tag_names,
-            field_names,
-            entries: vec![write_entry],
-        }
-    }
-
-    #[test]
-    fn test_build_schema_from_write_table_request() {
-        let schema_config = SchemaConfig {
-            default_timestamp_column_name: TIMESTAMP_COLUMN.to_string(),
-            ..SchemaConfig::default()
-        };
-        let write_table_request = generate_write_table_request();
-
-        let schema = build_schema_from_write_table_request(&schema_config, &write_table_request);
-        assert!(schema.is_ok());
-
-        let schema = schema.unwrap();
-
-        assert_eq!(8, schema.num_columns());
-        assert_eq!(2, schema.num_primary_key_columns());
-        assert_eq!(TIMESTAMP_COLUMN, schema.timestamp_name());
-        let tsid = schema.tsid_column();
-        assert!(tsid.is_some());
-
-        let key_columns = schema.key_columns();
-        assert_eq!(2, key_columns.len());
-        assert_eq!("tsid", key_columns[0].name);
-        assert_eq!(TIMESTAMP_COLUMN, key_columns[1].name);
-
-        let columns = schema.normal_columns();
-        assert_eq!(6, columns.len());
-
-        // sorted by column names because of btree
-        assert_eq!(FIELD1, columns[0].name);
-        assert!(!columns[0].is_tag);
-        assert_eq!(DatumKind::Double, columns[0].data_type);
-        assert_eq!(TAG1, columns[1].name);
-        assert!(columns[1].is_tag);
-        assert_eq!(DatumKind::String, columns[1].data_type);
-        assert_eq!(TAG2, columns[2].name);
-        assert!(columns[2].is_tag);
-        assert_eq!(DatumKind::String, columns[2].data_type);
-        assert_eq!(FIELD3, columns[3].name);
-        assert!(!columns[3].is_tag);
-        assert_eq!(DatumKind::String, columns[3].data_type);
-        assert_eq!(FIELD2, columns[4].name);
-        assert!(!columns[4].is_tag);
-        assert_eq!(DatumKind::Double, columns[4].data_type);
-        assert_eq!(FIELD4, columns[5].name);
-        assert!(!columns[5].is_tag);
-        assert_eq!(DatumKind::Boolean, columns[5].data_type);
-
-        for column in columns {
-            assert!(column.is_nullable);
-        }
-    }
-}
